+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Incorrect event end times [#97](https://github.com/rokwire/gateway-building-block/issues/97)
+
 [2.10.0] - 2024-05-16
 ### Changed
 - Handle location processing on WebTools import [#90](https://github.com/rokwire/gateway-building-block/issues/90)